Increase the verbosity of OpenBLAS's make build
Sadly the CIs are being temperamental and refuse to build something that
is too quiet. As a result the OpenBLAS `make` step is being terminated
prematurely. To workaround this issue, try increasing the verbosity of
the `make` step. It is worth noting that this has gone over the CI log
length limits historically. So this may not work either. However it's
simple and worth trying. If it works, we will have better log
information from the OpenBLAS build, which is a good thing.