Explicitly add `libgfortran-ng` to `host`
Builds on macOS will try to look for the `libgfortran` library during
the build. However as our compiler doesn't place that in the host
environment, macOS builds won't find it. This is demonstrated in the
linked log below. To fix this issue, just require `libgfortran-ng` be
installed in the host environment so that it can be linked to.
Admittedly this is a bit kludgy (though matches what we did previously);
however, this is something we have dealt with before when compiling on
macOS.

Linux can escape this problem as we provide `libgfortran` on the Docker
container. However to ensure it uses our `libgfortran`, go ahead and
include it there as well. This is what we did for Linux before. So it is
not too weird that we need to do it again.

ref: https://circleci.com/gh/conda-forge/openblas-feedstock/130