Fix OpenBLAS for Apple Silicon
Currently, OpenBLAS is broken on M1 machines and yields
incorrect results (see #139).

The fix for that problem will be included in the next release of
OpenBLAS, but it's unclear when the release will actually happen.

This means that currently, anyone using this version of OpenBLAS
on Apple Silicon may end up with incorrect calculations.

The patch is a simple one-liner.

Hence I propose to "backport" this patch here, and bump the build
number.

This change will only affect M1 users; behavior all other platforms
will be unaffected.